Improved mvc architecture
